(maint) Pin to Postgres 14.6
Postgres 14.7 started enforcing a restriction where databases could not
be created or dropped as part of a pipeline of SQL commands. We are
doing that in our `create-db` function when the user that is designated
to be the DB owner is not part of the `db-owner` role or a superuser.
That code should be fixed, though the fact that this isn't causing a
problem in PE, which is currently using Postgres 14.13, likely means
this path is unused.

In the meantime, this commit pins CI for this repo to 14.6, where the
existing code is still allowed.